No longer any need to get fetched vernacular pagenames in harvest_dag
